+++ core/classes/actions/class.action.editVhost.php
@@ -95,7 +95,7 @@ class ActionEditVhost
                     $neardWinbinder->resetProgressBar($this->wbProgressBar);
                     break;
                 }
-                if (file_put_contents($neardBs->getVhostsPath() . '/' . $serverName . '.conf', Util::getVhostContent($serverName, $documentRoot)) !== false) {
+                if (file_put_contents($neardBs->getVhostsPath() . '/' . $serverName . '.conf', $neardBins->getApache()->getVhostContent($serverName, $documentRoot)) !== false) {
                     $neardWinbinder->incrProgressBar($this->wbProgressBar);
                     
                     Util::addWindowsHost('127.0.0.1', $serverName);