+++ core/classes/tpls/app/class.tpl.app.status.php
@@ -21,9 +21,9 @@ class TplAppStatus
         
         //TODO: Manage services via Aestan or Win32Service ext ?
         return TplApp::getActionRun(Action::SWITCH_STATUS, array($status)) . PHP_EOL .
-            TplAppReload::getActionReload() . PHP_EOL .
             //TplAestan::getActionServiceRestart($neardBins->getApache()->getService()->getName());
             TplService::getActionRestart(BinApache::SERVICE_NAME) . PHP_EOL .
-            TplService::getActionRestart(BinXlight::SERVICE_NAME) . PHP_EOL;
+            TplService::getActionRestart(BinXlight::SERVICE_NAME) . PHP_EOL .
+            TplAppReload::getActionReload() . PHP_EOL;
     }
 }